+++ async-http-client/src/test/scala/org/http4s/client/asynchttpclient/AsyncHttpClientSpec.scala
@@ -1,8 +1,6 @@
-package org.http4s.client.asynchttpclient
-
-import org.http4s.client.ClientRouteTestBattery
-import org.http4s.util.threads.newDaemonPool
+package org.http4s
+package client
+package asynchttpclient
 
 class AsyncHttpClientSpec extends ClientRouteTestBattery("AsyncHttpClient",
-  AsyncHttpClient(bufferSize = 1,
-    customExecutor = Some(newDaemonPool("async-http-client-spec", timeout = true))))
+  AsyncHttpClient(bufferSize = 1)(Http4sSpec.TestExecutionContext))